fix log error and mail attach error - better mail code
