+++ .gitignore
@@ -147,5 +147,6 @@ GitHub.sublime-settings
 !.vscode/settings.json 
 !.vscode/tasks.json 
 !.vscode/launch.json 
+!.vscode/launch.json 
 !.vscode/extensions.json 
 .history
