+++ Backend/BackendApp/api/urls.py
@@ -6,11 +6,11 @@ from . import views
 
 
 urlpatterns = [
-    path('register/', views.register, name='register'),
-    path('login/', views.cust_login, name='login'),
-    path('home',views.home,name='home'),
-    path('adminhome',views.adminhome,name='adminhome'),
-    path('UpdateDonations',views.UpdateDonations,name='UpdatDonations'),
+    path('register/', views.register, name='register'),         # Register an new user
+    path('login/', views.cust_login, name='login'),             # login an user 
+    path('home',views.home,name='home'),                        # user page
+    path('adminhome',views.adminhome,name='adminhome'),         # adminpage
+    path('UpdateDonations',views.UpdateDonations,name='UpdatDonations'),   
     path('UpdateUsers',views.UpdateUsers,name='UpdateUsers'),
     path('resetpassword',views.resetpassword,name='ResetPassword'),
     path('DelDonoRec',views.DelDonaoRec,name='DelDonaoRec'),
